+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add new read-only attribute: `ipv4_cidr_block` to `ionoscloud_lan` resource and data source.
 - Make `volume` optional for `ionoscloud_server` resource
 - `name` attribute for `ionoscloud_auto_certificate` resource is now required.
+- Move postgresql to sdk-go-bundle.
 
 ## 6.6.9
 ### Features